ir/mir: Phase 6 wave 11.6 — split optimize.rs by pass (#252 Phase 6)
Pure refactor — zero behavior change. Splits the monolithic
2117-line `src/ir/mir/optimize.rs` into one module per
optimization pass under `src/ir/mir/optimize/`:

```
optimize/
  mod.rs              — module index + pipeline overview doc
  test_helpers.rs     — shared `span` / `one_fn_program` /
                        `body_of` fixtures (cfg(test) only)
  const_fold.rs       — wave 5 (literal arithmetic folding)
  dead_code.rs        — wave 6 (let-elision; exports
                        `pub(super) is_pure` for algebraic)
  inline.rs           — wave 7 (nullary literal inlining)
  algebraic.rs        — wave 8 (Int identity rewrites; reuses
                        `dead_code::is_pure`)
  bool_match.rs       — wave 9 (two-arm Bool match →
                        `IfThenElse`)
  branch_collapse.rs  — wave 10 (literal-cond IfThenElse
                        collapse)
```

Largest file now: 387 lines (`algebraic.rs` — pass body +
8 tests). All under the ~500 LOC review threshold.

## Why

The monolith had grown past the local "split when over 500
lines" guideline. Each pass owns its in-place walker, its
helper enums (Side / AlgReplace / BoolPat / BranchSide), and
its inline tests; mixing them in one file made it harder to
trace which helper belongs to which pass and risked
accidental cross-pass coupling. The split makes each pass
self-contained and reviewable in isolation.

## Cross-pass sharing

Only two helpers are intentionally shared:

- `dead_code::is_pure` — re-used by `algebraic` for the
  `x * 0` rewrite gate. Exported `pub(super)` and imported
  in `algebraic` via `use super::dead_code::is_pure;`.
- `test_helpers::{span, one_fn_program, body_of}` —
  per-pass test fixtures. Importing once at the module
  level keeps each pass's `mod tests` block focused on
  pass-specific assertions instead of boilerplate.

`const_fold::literal_eq` ended up not consumed by other
passes, so it stays `fn` (private) inside `const_fold`.

## Cascade tests still composable

End-to-end pipeline tests (`pipeline_compose_const_fold_algebraic_dce`,
`pipeline_bool_match_to_if_then_branch_collapse`, etc.) live
in whichever pass closes the cascade — the leaf pass imports
its upstream siblings from `super::`.

## Tests

- 1817 full-suite tests, 0 failures
- Zero new tests; same coverage as before the split
- Each `mod tests` block carries the tests for its own pass

## Test plan
- [x] cargo fmt + clippy clean
- [x] cargo test (1817 passing, 0 failed)

Follow-up na feedback drugiego AI o pliku > 500 linii.
